+++ src/main/java/snob/simulation/snob2/SonPartialView.java
@@ -130,7 +130,7 @@ public class SonPartialView extends AAgingPartialView {
                 }
                 // #B if it existed, keep the old age
                 if (found) {
-                    this.ages.add((Integer) removedAge.get(i));
+                    this.ages.add(removedAge.get(i));
                 } else {
                     // #C otherwise, it's a brand new one
                     this.ages.add(new Integer(0));
